Prevents EXPANDED_LENGTH from going out of bounds

+++ be/src/exprs/fpe.cpp
@@ -170,7 +170,7 @@ Status FPE::encrypt_num(std::string_view num_str, const std::vector<uint8_t>& ke
 
         std::string dec_int_part_str;
         dec_int_part_str = dec_part.substr(first_not_zero);
-        if (UNLIKELY(FPE::EXPANDED_LENGTH - first_not_zero <= 0)) {
+        if (UNLIKELY(FPE::EXPANDED_LENGTH - static_cast<int>(first_not_zero) <= 0)) {
             result.resize(result_len);
             value = std::move(result);
             return Status::OK();
